requirements.txt
numpy
scipy
scikit-learn
tqdm
transformers
sentence-transformers
pillow
requests
timm
pandas
clip
pycocotools
matplotlib
gradcam
opencv-python